+++ CHANGELOG.rst
@@ -8,6 +8,7 @@ Next version
 - Fixed the ordering calculation when cloning content: The gap for the new
   cloned content already had the correct size, but the base ordering value was
   incorrect.
+- Fixed the ordering calculation again.
 
 
 8.0 (2025-08-25)